Remove the special config file
To execute Puppet for a NetApp device it is necessary to add an entry
to the device configuration file. In the past you had to edit a second
`netapp.yml` file where you can drop the correct user credentials for
your NetApp device.

This does not seem necessary as the device configuration already
specifies an url attribute that can easily be used to store a username,
a password and the hostname of the device.

Use the URI instead and drop the special configuration file. This does
also make testing easier as you can overwrite the URI with a facts

<ADD EXAMPLE HERE!>